# Conflicts:
#	CHANGELOG.md
#	networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/BeaconChainMethods.java

+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/config/GetDepositContractTest.java
@@ -29,7 +29,7 @@ import tech.pegasys.teku.ethereum.execution.types.Eth1Address;
 
 class GetDepositContractTest extends AbstractMigratedBeaconHandlerTest {
   final Eth1Address eth1Address = dataStructureUtil.randomEth1Address();
-  final int chainId = spec.getGenesisSpecConfig().getDepositChainId();
+  final long chainId = spec.getGenesisSpecConfig().getDepositChainId();
   private final GetDepositContract.DepositContractData contractData =
       new GetDepositContract.DepositContractData(chainId, eth1Address);
